requirements.txt
anyio==3.5.0
asgiref==3.4.1
async-generator==1.10
attrs==21.4.0
beautifulsoup4==4.10.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.3
cryptography==36.0.2
Django==4.0.4
djangorestframework==3.13.1
fastapi==0.71.0
h11==0.12.0
idna==3.3
numpy==1.22.3
outcome==1.1.0
pandas==1.4.2
Pillow==9.1.1
psycopg2==2.9.3
pycparser==2.21
pydantic==1.9.0
PyJWT==2.3.0
pyOpenSSL==22.0.0
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.1
requests==2.27.1
selenium==4.1.3
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.1
SQLAlchemy==1.4.31
sqlparse==0.4.2
starlette==0.17.1
tk==0.1.0
trio==0.20.0
trio-websocket==0.9.2
twilio==7.8.2
typing_extensions==4.0.1
urllib3==1.26.9
uvicorn==0.17.0
wsproto==1.1.0
